S20-679, review fixes
