+++ pkg/api/operations/get_check_results_urlbuilder.go
@@ -33,7 +33,7 @@ func (o *GetCheckResultsURL) SetBasePath(bp string) {
 
 // Build a url path and query string
 func (o *GetCheckResultsURL) Build() (*url.URL, error) {
-	var result url.URL
+	var _result url.URL
 
 	var _path = "/checks"
 
@@ -41,9 +41,9 @@ func (o *GetCheckResultsURL) Build() (*url.URL, error) {
 	if _basePath == "" {
 		_basePath = "/api/v1"
 	}
-	result.Path = golangswaggerpaths.Join(_basePath, _path)
+	_result.Path = golangswaggerpaths.Join(_basePath, _path)
 
-	return &result, nil
+	return &_result, nil
 }
 
 // Must is a helper function to panic when the url builder returns an error
