+++ pkg/api/operations/get_plugin_config_urlbuilder.go
@@ -38,7 +38,7 @@ func (o *GetPluginConfigURL) SetBasePath(bp string) {
 
 // Build a url path and query string
 func (o *GetPluginConfigURL) Build() (*url.URL, error) {
-	var result url.URL
+	var _result url.URL
 
 	var _path = "/plugins/{pluginId}/config"
 
@@ -46,16 +46,16 @@ func (o *GetPluginConfigURL) Build() (*url.URL, error) {
 	if pluginID != "" {
 		_path = strings.Replace(_path, "{pluginId}", pluginID, -1)
 	} else {
-		return nil, errors.New("PluginID is required on GetPluginConfigURL")
+		return nil, errors.New("pluginId is required on GetPluginConfigURL")
 	}
 
 	_basePath := o._basePath
 	if _basePath == "" {
 		_basePath = "/api/v1"
 	}
-	result.Path = golangswaggerpaths.Join(_basePath, _path)
+	_result.Path = golangswaggerpaths.Join(_basePath, _path)
 
-	return &result, nil
+	return &_result, nil
 }
 
 // Must is a helper function to panic when the url builder returns an error
