+++ pkg/api/operations/get_plugins_urlbuilder.go
@@ -33,7 +33,7 @@ func (o *GetPluginsURL) SetBasePath(bp string) {
 
 // Build a url path and query string
 func (o *GetPluginsURL) Build() (*url.URL, error) {
-	var result url.URL
+	var _result url.URL
 
 	var _path = "/plugins"
 
@@ -41,9 +41,9 @@ func (o *GetPluginsURL) Build() (*url.URL, error) {
 	if _basePath == "" {
 		_basePath = "/api/v1"
 	}
-	result.Path = golangswaggerpaths.Join(_basePath, _path)
+	_result.Path = golangswaggerpaths.Join(_basePath, _path)
 
-	return &result, nil
+	return &_result, nil
 }
 
 // Must is a helper function to panic when the url builder returns an error
