+++ pkg/api/operations/get_promethius_integration_validations_responses.go
@@ -51,13 +51,13 @@ func (o *GetPromethiusIntegrationValidationsOK) WriteResponse(rw http.ResponseWr
 	rw.WriteHeader(200)
 	payload := o.Payload
 	if payload == nil {
+		// return empty array
 		payload = make([]*models.IntegrationComponent, 0, 50)
 	}
 
 	if err := producer.Produce(rw, payload); err != nil {
 		panic(err) // let the recovery middleware deal with this
 	}
-
 }
 
 /*GetPromethiusIntegrationValidationsDefault error
