+++ pkg/api/operations/get_promethius_integration_validations_urlbuilder.go
@@ -33,7 +33,7 @@ func (o *GetPromethiusIntegrationValidationsURL) SetBasePath(bp string) {
 
 // Build a url path and query string
 func (o *GetPromethiusIntegrationValidationsURL) Build() (*url.URL, error) {
-	var result url.URL
+	var _result url.URL
 
 	var _path = "/integration/prometheus/validations"
 
@@ -41,9 +41,9 @@ func (o *GetPromethiusIntegrationValidationsURL) Build() (*url.URL, error) {
 	if _basePath == "" {
 		_basePath = "/api/v1"
 	}
-	result.Path = golangswaggerpaths.Join(_basePath, _path)
+	_result.Path = golangswaggerpaths.Join(_basePath, _path)
 
-	return &result, nil
+	return &_result, nil
 }
 
 // Must is a helper function to panic when the url builder returns an error
