+++ pkg/api/server.go
@@ -187,6 +187,7 @@ func (s *Server) Serve() (err error) {
 
 		configureServer(domainSocket, "unix", string(s.SocketPath))
 
+		servers = append(servers, domainSocket)
 		wg.Add(1)
 		s.Logf("Serving analyze at unix://%s", s.SocketPath)
 		go func(l net.Listener) {
@@ -196,7 +197,6 @@ func (s *Server) Serve() (err error) {
 			}
 			s.Logf("Stopped serving analyze at unix://%s", s.SocketPath)
 		}(s.domainSocketL)
-		servers = append(servers, domainSocket)
 	}
 
 	if s.hasScheme(schemeHTTP) {
@@ -217,6 +217,7 @@ func (s *Server) Serve() (err error) {
 
 		configureServer(httpServer, "http", s.httpServerL.Addr().String())
 
+		servers = append(servers, httpServer)
 		wg.Add(1)
 		s.Logf("Serving analyze at http://%s", s.httpServerL.Addr())
 		go func(l net.Listener) {
@@ -226,7 +227,6 @@ func (s *Server) Serve() (err error) {
 			}
 			s.Logf("Stopped serving analyze at http://%s", l.Addr())
 		}(s.httpServerL)
-		servers = append(servers, httpServer)
 	}
 
 	if s.hasScheme(schemeHTTPS) {
@@ -313,6 +313,7 @@ func (s *Server) Serve() (err error) {
 
 		configureServer(httpsServer, "https", s.httpsServerL.Addr().String())
 
+		servers = append(servers, httpsServer)
 		wg.Add(1)
 		s.Logf("Serving analyze at https://%s", s.httpsServerL.Addr())
 		go func(l net.Listener) {
@@ -322,7 +323,6 @@ func (s *Server) Serve() (err error) {
 			}
 			s.Logf("Stopped serving analyze at https://%s", l.Addr())
 		}(tls.NewListener(s.httpsServerL, httpsServer.TLSConfig))
-		servers = append(servers, httpsServer)
 	}
 
 	wg.Wait()
